+++ service/gen.go
@@ -1,3 +1,3 @@
 package service
 
-//go:generate mockgen -package=servicetest -destination=servicetest/mocks.go github.com/abhinav/git-fu/service PR
+//go:generate mockgen -package=servicetest -destination=servicetest/mocks.go github.com/abhinav/git-fu/service PR,Review