+++ CHANGELOG.md
@@ -7,8 +7,9 @@ v0.5.0 (unreleased)
 -   `git pr rebase` now supports an `--only-mine` flag to rebase only pull
     requests owned by the current user.
 -   `git pr land` now verifies that a pull request passes all checks before
-    landing it. This includes verifying that all builds have succeeded and that
-    there are no blocking reviews. Use `--no-check` to skip this.
+    landing it. This includes verifying that all builds have succeeded, there
+    are no blocking reviews, and at least one user has approved the pull
+    request. Use `--no-check` to skip this check.
 
 
 v0.4.3 (2017-03-08)