+++ CHANGELOG.md
@@ -6,6 +6,8 @@ v0.5.0 (unreleased)
 
 -   `git pr rebase` now supports an `--only-mine` flag to rebase only pull
     requests owned by the current user.
+-   Fixed a bug where `git pr rebase` would not delete temporary branches if
+    the rebase failed.
 
 
 v0.4.3 (2017-03-08)